Подключение сервиса показа текущего курса
