+++ src/components/home/CurrencyRate/CurrencyRate.tsx
@@ -2,29 +2,29 @@ import { useDispatch, useSelector } from 'react-redux'
 import BankIcon from '../../../assets/icons/BankIcon'
 import { CurrencyItem } from '../../shared/CurrencyItem/CurrencyItem'
 import styles from './currency-rate.module.scss'
-import { fetchExchangeRate, selectExchangeRateFormatted } from '../../store/slices/ExchangeRateSlice/exchangeRateSlice'
 import { useEffect } from 'react'
 import type { AppDispatch } from '../../store/store'
+import { fetchExchangeRate } from '../../store/slices/exchangeRate/exchangedRateThunks'
+import { selectExchangeRateFormatted } from '../../store/slices/exchangeRate/exchangeRateSelectors'
 
 const CurrencyRate = () => {
-
   const dispatch = useDispatch<AppDispatch>()
-  const {data, formattedLastUpdate, status} = useSelector(selectExchangeRateFormatted)
+  const { data, formattedLastUpdate, status } = useSelector(selectExchangeRateFormatted)
 
   useEffect(() => {
-    if(status === 'idle') {
+    if (status === 'idle') {
       dispatch(fetchExchangeRate())
     }
   }, [dispatch, status])
 
   const displayCurrencies = ['USD', 'EUR', 'CNY', 'JPY', 'CAD', 'CHF']
 
-  const currencies = data 
-  ? displayCurrencies.map((code) => ({
-    code,
-    value: `${data[code]?.toFixed(2) ?? '-'} ₽`,
-  }))
-  : []
+  const currencies = data
+    ? displayCurrencies.map((code) => ({
+        code,
+        value: `${data[code]?.toFixed(2) ?? '-'} ₽`,
+      }))
+    : []
 
   return (
     <section className={styles['currency-rate']}>
@@ -32,7 +32,9 @@ const CurrencyRate = () => {
         <div className={styles['currency-rate__header']}>
           <p className={styles['currency-rate__title']}>Exchange rate in internet bank</p>
           <p className={styles['currency-rate__update-info']}>
-            <time>Update every 15 minutes {formattedLastUpdate ? `, ${formattedLastUpdate}` : ''}</time>
+            <time>
+              Update every 15 minutes {formattedLastUpdate ? `, ${formattedLastUpdate}` : ''}
+            </time>
           </p>
         </div>
