+++ src/components/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import exchangeRateReducer from './slices/ExchangeRateSlice/exchangeRateSlice'
+import exchangeRateReducer from './slices/exchangeRate/exchangeRateSlice'
 
 export const store = configureStore({
   reducer: {
